Confirm override of githook

+++ lib/quipper/rubocop/config.rb
@@ -1,23 +1,27 @@
-require "thor"
+# frozen_string_literal: true
 
+require "thor"
+require "pry"
+require "erb"
 
 module Quipper
   module Rubocop
     module Config
       class CLI < Thor
-
         RUBOCOP_CONFIG_FILE_NAME = ".rubocop.yml"
         RUBOCOP_TODO_FILE_NAME = ".rubocop_todo.yml"
         GITHOOK_FILE_PATH = ".git/hooks/pre-push"
 
         desc "install", "Install rubocop config"
         option :todo
-        option :hook
-        option :all
         def install
           create_rubocop_config!(RUBOCOP_CONFIG_FILE_NAME)
-          create_rubocop_todo_file! if options[:todo] || options[:all]
-          create_prepush_hook!(GITHOOK_FILE_PATH) if options[:hook] || options[:all]
+          create_rubocop_todo_file! if options[:todo]
+        end
+
+        desc "install_prepush", "Install prepush hook"
+        def install_prepush
+          create_prepush_hook!(GITHOOK_FILE_PATH)
         end
 
         desc "uninstall", "uninstall rubocop config"
@@ -31,12 +35,8 @@ module Quipper
 
         private
 
-        def template_path
-          File.expand_path("../../../templates", __dir__)
-        end
-
         def create_rubocop_config!(config_file_name)
-          puts "#{File.exist?(config_file_name) ? "overwrite" : "create"} #{config_file_name}"
+          puts "#{File.exist?(config_file_name) ? 'overwrite' : 'create'} #{config_file_name}"
           FileUtils.copy_file(File.join(template_path, config_file_name), config_file_name)
         end
 
@@ -46,26 +46,28 @@ module Quipper
         end
 
         def create_prepush_hook!(file_path)
-          puts "adding git prepush hook"
-          puts "#{File.exist?(file_path) ? "overwrite" : "create"} #{file_path}"
+          puts "adding git prepush hook at #{file_path}"
 
-          unless File.directory?(".git")
-            FileUtils.mkdir_p(".git")
-          end
-
-          unless File.directory?(".git/hooks")
-            FileUtils.mkdir_p(".git/hooks")
-          end
+          raise "File already exists at #{file_path}" if File.exist?(file_path)
 
           FileUtils.copy_file(File.join(template_path, ".githooks/pre-push"), file_path)
-          FileUtils.chmod(0755, file_path)
+          FileUtils.chmod(0o755, file_path)
+          puts "hook created!"
         end
 
         def remove_file(file_path)
-          if File.exist?(file_path)
+          if File.exist?(file_path) && added_by_this_gem?(file_path)
             File.delete(file_path)
             puts "Removing #{file_path}"
           end
+        end
+
+        def template_path
+          File.expand_path("../../../templates", __dir__)
+        end
+
+        def added_by_this_gem?(file_name)
+          File.read(file_name).include?("quipper-rubocop-config")
         end
       end
     end
